refactor: metrics in feature oriented architecture
