+++ src/lib/routes/client-api/index.ts
@@ -1,6 +1,6 @@
 import Controller from '../controller';
 import FeatureController from '../../features/client-feature-toggles/client-feature-toggle.controller';
-import MetricsController from '../../features/metrics/metrics';
+import MetricsController from '../../features/metrics/instance/metrics';
 import RegisterController from '../../features/metrics/instance/register';
 import { IUnleashConfig, IUnleashServices } from '../../types';
